moved debugging code to test
