+++ openpnm/utils/Settings.py
@@ -121,124 +121,3 @@ class SettingsAttr:
     @property
     def _attrs(self):
         return self.__dir__()
-
-
-if __name__ == "__main__":
-
-    from traits.api import Int, List, Str, Float, TraitError, ListStr
-
-    # %% Standard initialization
-    class S1(SettingsData):
-        r"""
-        This is a docstring
-        """
-        a = Int(1)
-        b = Int(2)
-        d = List(Str())
-
-    sets1 = SettingsAttr(S1())
-    print(sets1)
-    assert "This is a docstring" in sets1.__doc__
-
-    # %% Dataclass-style and testing inferred type is enforced
-    class S2:
-        r"""
-        This is a docstring
-        """
-        a = 2
-        b = 3
-
-    sets2 = SettingsAttr(S2())
-    print(sets2)
-    assert "This is a docstring" in sets2.__doc__
-
-    # %% Inheritance and testing immutability
-
-    class S3(S2):
-        r"""
-        Different docstring
-        """
-        b = 3
-        c = 4
-
-    sets3 = SettingsAttr(S3())
-    sets3.b = 44
-    print(sets3)
-    assert dir(sets3) == ['a', 'b', 'c']
-    assert sets2.b != sets3.b
-    assert "Different docstring" in sets3.__doc__
-
-    try:
-        sets3.b = 'nope'
-    except TraitError as e:
-        print(e)
-
-    # %% Adding new attr and testing type is enforced
-    class S4:
-        r"""
-        This is a docstring
-        """
-        a = 2
-        b = 3.5
-
-    sets4 = SettingsAttr(S4())
-    sets4.c = Float(1.5)
-    print(sets4)
-    try:
-        sets4.c = "string"
-    except TraitError as e:
-        print(e)
-
-    # %% From dict
-    d = {'a': 3, 'b': 4.5, 'c': []}
-
-    sets5 = SettingsAttr(d)
-    try:
-        sets5.c = "string"
-    except TraitError as e:
-        print(e)
-    sets5.c.append(1)
-    print(sets5)
-
-    # %% Update from dict
-    class S6(SettingsData):
-        r"""
-        This is a docstring
-        """
-        a = Int(1)
-        b = Int(2)
-
-    sets6 = SettingsAttr(S6())
-    assert sets6.a == 1
-    sets6._update({'a': 22, 'c': 5.5})
-    assert sets6.a == 22
-    print(sets6)
-
-    # %% Update from dataclass
-    class S7(SettingsData):
-        r"""
-        This is a docstring
-        """
-        a = Int(1)
-        b = Int(2)
-
-    sets7 = SettingsAttr(S7())
-    assert sets7.a == 1
-    class Data:
-        a = 22
-        c = 5.5
-    sets7._update(Data())
-    print(sets7)
-    assert sets7.a == 22
-
-
-
-
-
-
-
-
-
-
-
-
